[utils][m]: Seperate utility functions into seperate folder
